updated backend

+++ back-end/db/schema.sql
@@ -12,7 +12,7 @@ CREATE TABLE
 products (
     id serial PRIMARY KEY,
     name TEXT NOT NULL,
-    price NUMERIC NOT NULL,
+    price INTEGER NOT NULL,
     description TEXT NOT NULL,
     image  TEXT NOT NULL,
     type TEXT NOT NULL
